requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
numpy>=1.19.0

# Audio Processing
librosa>=0.10.0

# Subtitle Processing
ass>=0.5.0

# Configuration Management
PyYAML>=6.0

# Dataset Storage
h5py>=3.0.0

# Signal Processing and Scientific Computing
scipy>=1.7.0

# Training Monitoring and Logging
tensorboard>=2.8.0

# Optional: For better performance and compatibility
# Uncomment these if you encounter issues or want optimal performance

# For CUDA users (faster audio processing)
# numba>=0.56.0

# For better audio format support
# soundfile>=0.12.0

# For visualization and debugging (optional)
# matplotlib>=3.5.0
# seaborn>=0.11.0

# Development and testing (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=4.0.0